+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `Package::signature_key_fingerprints()`
+
 ### Changed
 
 - Bump `pgp` to `0.19.0`